nftvisionbackend/requirements.txt
argon2-cffi==20.1.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
decorator==5.0.9
defusedxml==0.7.1
Django==2.2.19
entrypoints==0.3
fpdf==1.7.2
glfw==2.1.0
gunicorn==20.1.0
idna==2.10
image==1.5.33
importlib-metadata==2.1.1
jedi==0.17.2
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==1.1.1
mistune==0.8.4
multipledispatch==0.6.0
nbconvert==5.6.1
nbformat==5.1.3
packaging==20.9
pandocfilters==1.4.3
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
prometheus-client==0.11.0
prompt-toolkit==2.0.10
ptyprocess==0.7.0
pycairo==1.19.1
pycparser==2.20
pyparsing==2.4.7
pyrr==0.10.3
pyrsistent==0.17.3
python-dateutil==2.8.2
pytz==2021.1
pyzmq==20.0.0
qtconsole==4.7.7
QtPy==1.9.0
requests==2.25.1
Send2Trash==1.8.0
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
terminado==0.8.3
testpath==0.5.0
text-to-image==0.0.5
tornado==6.1
traitlets==4.3.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
whitenoise
zipp==1.2.0
